acme: Avoid reset of Client.net to None (#5738)
The issue: For net=None self.net was first set to
a new instance of ClientNetwork but then the original
net with value None is passed to ClientBase.__init__
which resets self.net to None.

The fix: Client.__init__ does no longer touch self.net
but leaves that to ClientBase.__init__.  Local parameter
net is modified before passing it on, instead.