Apache: list everywhere for challenges (#6046)
A site can have multiple hosts sharing a certificate, e.g.:
<VirtualHost localhost:443>
ServerName x.example.com
</VirtualHost>
<VirtualHost x.example.com:443>
ServerName x.example.com
</VirtualHost>
<VirtualHost 10.0.0.1:443>
ServerName x.example.com
</VirtualHost>

The goal of the SNI challenge is to answer requests for specially defined servernames.
We really do not care how they are delivered to apache, as long as we can get apache to
return our specially crafted certificate.

As such, the simplest approach seems to be to just use every possible vhost.

We could take some effort to filter for the domains we suspect are relevant,
but it really isn't worth it. The only client that wants to talk to our
strange domain is ACME, and we'll let apache filter on that based on our
server name.

There are some potentially bizarre ways for traffic to actually come into localhost
(I actually have done this for other systems), but in general, it is more likely that
it will come in via some other address. That said, there does not appear to be any
harm/risk/reason not to just listen for any possible entrypoint.

It is definitely better to listen to all possible entrypoints than it is to only listen
to localhost, which will on average never be able to answer the request.