Clarify that configurations can be invalid
A configuration file will typically point to files...
If those files are supposed to be symbolic links but are not,
then the configuration is invalid.

Any normal person upon being told that a configuration file is
invalid would compare the configuration file to another
configuration file, and not spot anything particularly wrong with it.

The problem in that case is the configuration, which includes the
configuration file and everything it references.
-- NOT the configuration file itself.

#6276