Implement end-to-end tests for cloudflare plugin
Implement BATS tests for the dns-cloudflare plugin to verify obtaining
and renewing a certificate, obtaining a certificate for multiple
domains, and a variety of negative cases involving incorrect, invalid,
or missing credentials.

These test cases are skipped if the appropriate environment variables
are not set. This will allow for execution of the subset of the end-to-
end tests for DNS plugins for which credentials are available.